UI Changes
- major ui changes and Seprate file creation done.
