+++ TaskManager-SwiftUI/TaskManager-SwiftUI/View/AddTaskButton.swift
@@ -1,0 +1,27 @@
+import SwiftUI
+
+struct AddTaskButton: View {
+    @Binding var showingAddTask: Bool
+    
+    var body: some View {
+        VStack {
+            Spacer()
+            HStack {
+                Spacer()
+                Button(action: { showingAddTask = true }) {
+                    Image(systemName: "plus")
+                        .font(.title2)
+                        .fontWeight(.semibold)
+                        .foregroundColor(.white)
+                        .frame(width: 56, height: 56)
+                        .background(GradientUtility.buttonGradient)
+                        .clipShape(Circle())
+                        .shadow(color: Color.blue.opacity(0.3), radius: 10, x: 0, y: 5)
+                }
+                .padding(.trailing, 20)
+                .padding(.bottom, 20)
+            }
+        }
+    }
+}
+
