+++ TaskManager-SwiftUI/TaskManager-SwiftUI/View/CircularProgressRing.swift
@@ -1,0 +1,27 @@
+import SwiftUI
+
+struct CircularProgressRing: View {
+    let progress: Double
+    let size: CGFloat
+    
+    var body: some View {
+        ZStack {
+            Circle()
+                .stroke(Color.gray.opacity(0.2), lineWidth: 4)
+            
+            Circle()
+                .trim(from: 0, to: progress)
+                .stroke(Color.blue, style: StrokeStyle(
+                    lineWidth: 4,
+                    lineCap: .round
+                ))
+                .rotationEffect(.degrees(-90))
+            
+            Text("\(Int(progress * 100))%")
+                .font(.system(size: size * 0.3))
+                .fontWeight(.medium)
+        }
+        .frame(width: size, height: size)
+    }
+}
+
